+++ CHANGELOG.md
@@ -14,7 +14,7 @@ The **need for configuration updates** is **marked bold**.
 ### Added
 
 - Added base entities and repositories for Aggregated Supply Chain Data ([#1182](https://github.com/eclipse-tractusx/puris/pull/1182))
-- Added mapper service for Aggregated Supply Chain Data ([#1184](https://github.com/eclipse-tractusx/puris/pull/1184))
+- Added mapper service and test for Aggregated Supply Chain Data ([#1184](https://github.com/eclipse-tractusx/puris/pull/1184))
 
 ### Changed